Turn `rb_classext_t.fields` into a T_OBJECT
This aims to solve two problems.

First, it solves the problem of namspaced classes having
a single `shape_id`. Now each namespaced classext
has a `T_OBJECT` that can hold the namespace specific
shape.

Second, it open the door to later make class instance variable
writes atomics, hence be able to read class variables
without locking the VM.
In the future, in multi-ractor mode, we can do the write
on a copy of the `fields_obj` and then atomically swap it.

Considerations:

  - T_OBJECT isn't ideal, it probably should be a `T_IMEMO`
    but this require a very large refatoring.
  - Right now the `RClass` shape_id is always synchronized,
    but with namespace we should likely mark classes that have
    multiple namespace with `INVALID_SHAPE`.